add-equity - save

+++ src/main/kotlin/com/github/vantonov1/meet/entities/District.kt
@@ -4,4 +4,4 @@ import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Table
 
 @Table
-data class District(@Id val id: Byte? = null, val name: String, val city: Short)
+data class District(@Id val id: Short? = null, val name: String, val city: Short)
